a function that searches for a value in a sorted array of integers using the Exponential search algorithm

+++ 0x1E-search_algorithms/103-exponential.c
@@ -84,7 +84,7 @@ int exponential_search(int *array, size_t size, int value)
 
 	while (array[index] < value && index < size)
 	{
-		printf("Value checked array[%d] = [%d]\n", (int)index, array[index]);
+		printf("Value checked array[%lu] = [%d]\n", index, array[index]);
 		index *= 2;
 	}
